api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.10
pydantic==2.6.4
python-dotenv==1.0.1
